+++ tests/managers/test_kubernetes.py
@@ -6,7 +6,3 @@ class Test_Kubernetes(unittest.TestCase):
 
     def setUp(self) -> None:
         self.manager = Kubernetes()
-
-    # def test_list_pod(self):
-    #     self.manager.ls()
-    #     self.fail()